Adding Unreal .gitignore
Will prevent cluttering and LFS files on the GitHub repo that are not needed. This will ensure that the following folders are not uploaded to the GitHub repo:
1. Intermediate
2. Derived Data Cache 

And other compiled binaries that can be regenerated on individual systems for development. Keeps the size small, simultaneously maintaining less file count.